requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-limiter==3.8.0
PyJWT==2.8.0
bcrypt==4.1.3
firebase-admin==6.5.0
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
gunicorn==21.2.0
